Split the commands module into more smaller ones
